Restore recap worker panic parity
